lib: rename _component to _molecule
